refactor(layout): improve SidebarLayout structure and maintainability
